Day 22, Part 1: Get small and medium tests working for part 1
This still does not work for the full part 1 input, and I think I've
identified some areas to look for bugs. I think I'd like to convert all
the storage of cubes to sets so that there are no duplicates, and so
that it's easier to remove by value rather than doing gymnastics with
indicies to remove in the proper order, etc.