Day 23, Part 1: Add next_moves to optimize from possible_moves
If any possible move would move into a home room, then just return that
move, because we don't care about any of the other moves. This helps. It
gets the run time for the input down to the 10-20s range instead of 90s,
but I think it's missing the case where there are N possible moves to a
home room, in which case we want to just do all of them in one
iteration. We also want to check to see if a burrow is organized after
applying the moves, rather than putting it back into the priority queue
and waiting for it to come up again. So, still some TODOs.