Day 23, Part 1: Return an iterator from room_for
I'm not really sure if this is a real optimization, because the complier
would possibly/probably inline the results of room_for for each type,
but since we're turning the result into an interator anyway, I decided
to just have it return the iterator directly and then the functions that
use it can use the iterator directly, without having to create the Vec
and copy it around. Could be an improper optimization, but just wanted
to give it a shot.